- 调整代码

+++ Engine/Shader/Shader/TransferSceneData.glsl
@@ -17,6 +17,6 @@ USE_TRANSFER_SCENE_DATA_INDEX_STORAGE;
 
 void main()
 {
-	uint indexWork = uint(gl_GlobalInvocationID.x);
+	uint indexWork = gl_GlobalInvocationID.x;
 	sceneData.data[transferSceneDataIndex.index[indexWork]] = transferSceneData.data[indexWork];
 }
